requirements.txt
streamlit==1.26.0
PyPDF2==3.0.1
python-dotenv==1.0.0
google-generativeai==0.3.0
